+++ consensus-highlighter.py
@@ -3,9 +3,9 @@
 
 import sys
 
-__version__: str  = '1.0.c'
+__version__: str  = '1.0.d'
 # Year, month, day
-__last_update_date__: str = '2021-06-10'
+__last_update_date__: str = '2021-06-11'
 __min_python_version__: float = 3.6
 __author__ = 'Maxim Sikolenko'